test: migrate tests about auth headers to KintoneRequestConfigBuilder.test
