+++ packages/rest-api-client/src/__tests__/KintoneRequestConfigBuilder.test.ts
@@ -1,6 +1,9 @@
 import { KintoneRequestConfigBuilder } from "../KintoneRequestConfigBuilder";
 import FormData from "form-data";
 
+import { injectPlatformDeps } from "../platform";
+import * as nodeDeps from "../platform/node";
+
 describe("KintoneRequestConfigBuilder", () => {
   const baseUrl = "https://example.kintone.com";
   const headers = {
@@ -179,5 +182,140 @@ describe("options", () => {
       { key: "value" }
     );
     expect(requestConfig).toHaveProperty("httpsAgent");
+  });
+});
+
+describe("Headers", () => {
+  const baseUrl = "https://example.com";
+
+  it("Password auth", () => {
+    const USERNAME = "user";
+    const PASSWORD = "password";
+    const auth = {
+      username: USERNAME,
+      password: PASSWORD,
+    };
+    const kintoneRequestConfigBuilder = new KintoneRequestConfigBuilder({
+      baseUrl,
+      auth,
+    });
+    expect(
+      kintoneRequestConfigBuilder.build("get", "/k/v1/record.json", {}).headers
+    ).toEqual({
+      "X-Cybozu-Authorization": Base64.encode(`${USERNAME}:${PASSWORD}`),
+    });
+  });
+
+  it("ApiToken auth", () => {
+    const API_TOKEN = "ApiToken";
+    const auth = {
+      apiToken: API_TOKEN,
+    };
+    const kintoneRequestConfigBuilder = new KintoneRequestConfigBuilder({
+      baseUrl,
+      auth,
+    });
+    expect(
+      kintoneRequestConfigBuilder.build("get", "/k/v1/record.json", {}).headers
+    ).toEqual({
+      "X-Cybozu-API-Token": API_TOKEN,
+    });
+  });
+
+  it("ApiToken auth using multiple tokens as comma-separated string", () => {
+    const API_TOKEN1 = "ApiToken1";
+    const API_TOKEN2 = "ApiToken2";
+    const auth = {
+      apiToken: `${API_TOKEN1},${API_TOKEN2}`,
+    };
+    const kintoneRequestConfigBuilder = new KintoneRequestConfigBuilder({
+      baseUrl,
+      auth,
+    });
+    expect(
+      kintoneRequestConfigBuilder.build("get", "/k/v1/record.json", {}).headers
+    ).toEqual({
+      "X-Cybozu-API-Token": `${API_TOKEN1},${API_TOKEN2}`,
+    });
+  });
+
+  it("ApiToken auth using multiple tokens as array", () => {
+    const API_TOKEN1 = "ApiToken1";
+    const API_TOKEN2 = "ApiToken2";
+    const auth = {
+      apiToken: [API_TOKEN1, API_TOKEN2],
+    };
+    const kintoneRequestConfigBuilder = new KintoneRequestConfigBuilder({
+      baseUrl,
+      auth,
+    });
+    expect(
+      kintoneRequestConfigBuilder.build("get", "/k/v1/record.json", {}).headers
+    ).toEqual({
+      "X-Cybozu-API-Token": `${API_TOKEN1},${API_TOKEN2}`,
+    });
+  });
+
+  it("Session auth", () => {
+    const auth = {};
+    const kintoneRequestConfigBuilder = new KintoneRequestConfigBuilder({
+      baseUrl,
+      auth,
+    });
+    expect(
+      kintoneRequestConfigBuilder.build("get", "/k/v1/record.json", {}).headers
+    ).toEqual({
+      "X-Requested-With": "XMLHttpRequest",
+    });
+  });
+
+  it("should use Session auth if auth param is not specified", () => {
+    const kintoneRequestConfigBuilder = new KintoneRequestConfigBuilder({
+      baseUrl,
+    });
+    expect(
+      kintoneRequestConfigBuilder.build("get", "/k/v1/record.json", {}).headers
+    ).toEqual({
+      "X-Requested-With": "XMLHttpRequest",
+    });
+  });
+
+  it.skip("should raise an error if trying to use session auth in Node.js environment", () => {
+    injectPlatformDeps(nodeDeps);
+    const kintoneRequestConfigBuilder = new KintoneRequestConfigBuilder({
+      baseUrl,
+    });
+    expect(
+      kintoneRequestConfigBuilder.build("get", "/k/v1/record.json", {})
+    ).toThrow(
+      "session authentication is not supported in Node.js environment."
+    );
+  });
+
+  it("OAuth token auth", () => {
+    const auth = { oAuthToken: "oauth-token" };
+    const kintoneRequestConfigBuilder = new KintoneRequestConfigBuilder({
+      baseUrl,
+      auth,
+    });
+    expect(
+      kintoneRequestConfigBuilder.build("get", "/k/v1/record.json", {}).headers
+    ).toEqual({
+      Authorization: `Bearer ${auth.oAuthToken}`,
+    });
+  });
+
+  it("Basic auth", () => {
+    const basicAuth = { username: "user", password: "password" };
+    const kintoneRequestConfigBuilder = new KintoneRequestConfigBuilder({
+      baseUrl,
+      basicAuth,
+    });
+    expect(
+      kintoneRequestConfigBuilder.build("get", "/k/v1/record.json", {}).headers
+    ).toEqual({
+      Authorization: `Basic ${Base64.encode("user:password")}`,
+      "X-Requested-With": "XMLHttpRequest",
+    });
   });
 });
